requirements.txt
#pip install -r requirements.txt
openai-whisper
ollama
chromadb
moviepy
torch
numpy<1.24
fastapi
inaSpeechSegmenter
pandas
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
yt-dlp
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv